# Conflicts:
#	res/menu/throttle_menu.xml
#	src/jmri/enginedriver/threaded_application.java

+++ src/jmri/enginedriver/ConsistEdit.java
@@ -232,7 +232,7 @@ public class ConsistEdit extends Activity implements OnGestureListener {
 
                 if (!consist.getLeadAddr().equals(addr)) {
                     consist.remove(addr);
-                    mainapp.sendMsg(mainapp.comm_msg_handler, message_type.RELEASE, addr, whichThrottle);   //release the loco
+                    mainapp.sendMsg(mainapp.comm_msg_handler, message_type.RELEASE, addr, (int) whichThrottle);   //release the loco
                     refreshConsistLists();
                 }
                 return true;
